Exlude data from IPUMS NHGIS
This way I comply with their terms of use. I've given up on using the census bureau APIs -- not because they're hard to use, but because they're only available out to 1990. With this dataset, I could go all the way back to the 1770s, if I'd wanted to (though climate data restricts that).