Fetch Population Info from 1970 to 2010
These changes allow for the fetching of the total population of each metro area, going back to 1970.

Because there's not a convienent source for population data by metro, I first use a dileaniation file to convert each metro area into a list of it's constituent counties.

Because the census bureau API is only good back through 1990 and has many inconsistencies that make it challenging to work with, I'm using decennial census data from IPUMS NGHIS, which has consolidated all of the poulation data for all counties in the country from 1770 to 2010.

After we get the populations, we proceed to get the stations as previously.