Attempt to Eliminate Global Warming Effect
Try minimizing the effect of global warming. To do this, I estimate the efect of global warming using the most rural 15% of stations. Because they're still effected by the UHI, despite being rural, I apply a correction for both the metro population and the distance from the metro center it is.

I take the average of all of these by decade to create a correction factor. This should clear away most of the time-dependency of the data that's based on AGW as opposed to pouplation growth.